Move the helper code for the parameterized test to separate class
